+++ import_gr2_for_blender5.py
@@ -1,47 +1,25 @@
 # ADDON VERSION
 
-# Collada import replacement for blender 5.0
-# Not perfect, but at present it does successfully import GR2 and DAE, with a start on bone alignment. 
-# Still needs more work but it does basically do what I need it to. Meshes + armatures are functional. Animations import, but obviously the bone misalignment is most apparent there.
+#changing the name: GR2 Importer 2025, instead of 'GR2 Importer for Blender 5'. 
+# #Shouldn't be too hard to make it backwards compatible with 4.3~, so a year just to indicate 'this is modern' feels better than a specific blender release.
 
 # No export functionality at all yet, may implement it later but I only need import for myself so that's my focus.
- 
 #  -- harpoon
   
 # 22/10/2025
-
 
 armaturepath = None # here to make sure there's always something. I kept commenting it out in testing.
 
 import_list = [r"F:\BG3 Extract PAKs\PAKs\Models\Generated\Public\Shared\Assets\Characters\_Anims\_Creatures\Intellect_Devourer\Resources\Proxy_INTDEV_A.GR2", r"F:\test\gltf_tests\rpremixed anim and skel as skel to mesh.gr2", r"F:\test\gltf_tests\randompeaceintdev defaults copyskel.gr2", r"F:\BG3 Extract PAKs\PAKs\Models\Generated\Public\Shared\Assets\Characters\_Models\_Creatures\Intellect_Devourer\Resources\INTDEV_CIN.GR2", r"F:\test\gltf_tests\random_peace_to_intdev_then_intdev_cin_merged_with_the_combined_copy_skel.gr2", r"F:\test\gltf_tests\random_peace_to_intdev_then_intdev_cin_merged_with_the_combined.gr2", r"F:\test\gltf_tests\random_peace_with_intdev_cin.gr2", 
 r"F:\BG3 Extract PAKs\PAKs\Models\Public\Shared\Assets\Characters\_Anims\_Creatures\OwlBear\OWLBEAR_Cub_Rig\_Construction\Owlbear_Cub_Rig_DFLT_CINE_Curious_HeadTilt_01.GR2", r"F:\BG3 Extract PAKs\PAKs\Models\Public\Shared\Assets\Characters\_Anims\_Creatures\OwlBear\OWLBEAR_Cub_Base.GR2", r"F:\test\gltf_tests\owlbear_mesh.dae", r"F:\BG3 Extract PAKs\PAKs\Models\Public\Shared\Assets\Characters\_Anims\_Creatures\Intellect_Devourer\INTDEV_Rig\INTDEV_Rig_DFLT_IDLE_Random_Peace_01.GR2", r"F:\BG3 Extract PAKs\PAKs\Models\Public\Shared\Assets\Characters\_Anims\_Creatures\Intellect_Devourer\INTDEV_Base.GR2", r"F:\BG3 Extract PAKs\PAKs\Models\Generated\Public\Shared\Assets\Characters\_Models\_Creatures\Intellect_Devourer\Resources\INTDEV_CIN.GR2"]
 
-#file_to_import = r"F:\BG3 Extract PAKs\PAKs\Models\Generated\Public\Shared\Assets\Characters\_Anims\_Creatures\Intellect_Devourer\Resources\Proxy_INTDEV_A.GR2"
-#file_to_import = r"F:\test\glb_tests\rpremixed anim and skel as skel to mesh.gr2" # the name lies, it's just intdev cin.
-#file_to_import = r"F:\test\glb_tests\randompeaceintdev defaults copyskel.gr2" ## skel +  animation, no mesh
-## file_to_import = r"F:\BG3 Extract PAKs\PAKs\Models\Generated\Public\Shared\Assets\Characters\_Models\_Creatures\Intellect_Devourer\Resources\INTDEV_CIN.GR2"
-#armaturepath = r"F:\BG3 Extract PAKs\PAKs\Models\Generated\Public\Shared\Assets\Characters\_Models\_Creatures\Intellect_Devourer\Resources\INTDEV_CIN.GR2"
-
-#file_to_import = r"F:\test\glb_tests\random_peace_to_intdev_then_intdev_cin_merged_with_the_combined_copy_skel.gr2"
-#file_to_import = r"F:\test\glb_tests\random_peace_to_intdev_then_intdev_cin_merged_with_the_combined.gr2" # skel + anim, no mesh, made in stage 2 of testing.
-#file_to_import = r"F:\test\glb_tests\random_peace_with_intdev_cin.gr2" # done in converterapp, random peace as input file, intdev cin as armature. copy skeleton from and copy skeleton both on. Everyting else at defaults.
-#file_to_import = r"F:\BG3 Extract PAKs\PAKs\Models\Public\Shared\Assets\Characters\_Anims\_Creatures\OwlBear\OWLBEAR_Cub_Rig\_Construction\Owlbear_Cub_Rig_DFLT_CINE_Curious_HeadTilt_01.GR2" ## head tilt animation
-#armaturepath = r"F:\BG3 Extract PAKs\PAKs\Models\Public\Shared\Assets\Characters\_Anims\_Creatures\OwlBear\OWLBEAR_Cub_Base.GR2"
-
-#file_to_import = r"F:\test\glb_tests\owlbear_mesh.dae"
-
-#file_to_import = r"F:\BG3 Extract PAKs\PAKs\Models\Public\Shared\Assets\Characters\_Anims\_Creatures\Intellect_Devourer\INTDEV_Rig\INTDEV_Rig_DFLT_IDLE_Random_Peace_01.GR2"
-#armaturepath =
 file_to_import = r"F:\BG3 Extract PAKs\PAKs\Models\Public\Shared\Assets\Characters\_Anims\_Creatures\Intellect_Devourer\INTDEV_Base.GR2"
 
-inputs = None, None#armaturepath, file_to_import
+#inputs = None, None#armaturepath, file_to_import
 
-version = "0.58" # metadata checker now works entirely within the addon, with a popup with the results (as well as print to terminal)
+version = "0.61" # just some cleanup. 
 
-mode = "metadata_only"#"all"
-#mode = "mass_metadata" #"all"#"metadata only"  # other options: "all", "metadata only", "mass_metadata"
-metadata = True
-specified_collection = "bulk_anim_test_1"
+# Turns out the retargeter is broken in Blender v5, so that's hurt my soul a little. 
 
 json_output = False
 jsonout_file = f"F:\\test\\gr2_metadata.json" #None
@@ -53,11 +31,8 @@ import subprocess
 import tempfile
 from pathlib import Path
 
-use_existing_collection = True
-custom_bones_on = True
-
 divineexe = r"F:\Blender\Addons etc\Packed\Tools\Divine.exe"
-divinedir = r"F:\Blender\Addons etc\Packed\Tools"
+divinedir = str(Path(divineexe).parent)
 rootreader = r"D:\Git_Repos\GR2_importer_for_blender_5\rootreader\bin\Debug\net8.0\rootreader.exe"
 
 status_definitions = {
@@ -75,45 +50,57 @@ status_definitions = {
     7: "Mesh, armature & animation"
 }
 
-has_armature = [1, 3, 5, 7]
+has_armature = [1,3,5,7]
+null_status = [0, 00, None, ""]
+
+def initial_setup(mode, settings):
+
+    print(f"Settings: {settings}")
+    if mode == "import":
+        path = divineexe#"addon preferences/divine" # these don't exist yet.
+    elif mode == "metadata":
+        path = rootreader#"addon preferences/metadata_checker"
 
 #--- Check required files exist. --- #
+    def check_for_exe(path):
+        if not Path(path).is_file():
+            print(f"Required file not found: {path}. Please check the paths in the script.") 
+            return list(f"Cannot initialise, required file not found: `{path}`.", "Please add correct path in Add-on Preferences.")
 
-for path in [divineexe, rootreader]:
-    if not Path(path).is_file():
-        print(f"Required file not found: {path}. Please check the paths in the script.")
-        print("Press 'Enter' to continue, type 'Exit' to exit.")
-        user_input = input()
-        if user_input.lower() == 'exit':
-            exit()
+    check_for_exe(path)
 
 ### --- Ensure console is visible before running anything. --- ###
-def is_console_visible():
-    import ctypes
-    hwnd = ctypes.windll.kernel32.GetConsoleWindow()
-    return hwnd != 0 and ctypes.windll.user32.IsWindowVisible(hwnd)
-    
-#if not is_console_visible():
-#    bpy.ops.wm.console_toggle()
+    def is_console_visible():
+        import ctypes
+        hwnd = ctypes.windll.kernel32.GetConsoleWindow()
+        return hwnd != 0 and ctypes.windll.user32.IsWindowVisible(hwnd)
 
-# Make sure the native glb importer is enabled for later. # 
-default, enabled = check("io_scene_gltf2")
-if not enabled:
-    try:
-        enable("io_scene_gltf2", default_set=True, persistent=True)
-    except Exception as e:
-        print(f"Failed to enable gltf importer: {e}")
-        print("Please enable the 'gltf 2.0 format' addon manually in Blender preferences and try again.")
-        exit()
+    if settings.get("open_console"):  
+        if not is_console_visible():
+            bpy.ops.wm.console_toggle()
 
+    if mode == "import":
+        default, enabled = check("io_scene_gltf2")
+        if not enabled:
+            try:
+                enable("io_scene_gltf2", default_set=True, persistent=True)
+            except Exception as e:
+                print(f"Failed to enable glTF native importer: {e}")
+                print("Please enable the 'glTF 2.0 Format' addon manually in Blender preferences and try again.")
+                return list(f"Failed to enable glTF Importer: {e}", "Please enable `glTF 2.0 Format` addon manually in Blender Add-ons and try again.")
+
+## === Little utilities ===
 def print_me(status, *args, **kwargs):
     if status:
         print(*args, **kwargs)
 
 def get_filename_ext(filepath = None):
     if not filepath:
-        return None, None
+        return None, None, None
     
+    print(filepath)
+    if "\\" not in filepath:
+        return None, None, None
     directory, filename = str(filepath).rsplit("\\", 1)
     ext = filename.split(".")[-1]
     return directory, filename, ext.lower()
@@ -126,11 +113,11 @@ def get_temppath():
 
 def check_file_exists(origin, filepath):
 
-    if filepath == None or not Path(filepath).is_file(): # test this.
+    if filepath == None or not Path(filepath).is_file():
         print(f"No file provided for {origin}.")
         return False
-
     return True
+
 
 def get_metadata(filepath, printme):
 
@@ -139,118 +126,138 @@ def get_metadata(filepath, printme):
     result = subprocess.run(
         [rootreader, filepath],
         capture_output=True,
-        text=True, cwd=divinedir ### `cwd`:  set to divine dir to ensure any dependent files are found.
-    )
+        text=True, cwd=divinedir) ### `cwd`:  set to divine dir to ensure any dependent files are found.
+
+    print(f"Ran subprocess on {filepath}")
 
     meshes, armatures, animations = 0, 0, 0
     if result.returncode == 0:
         root_data = json.loads(result.stdout)
-        if "Skeletons" in root_data and root_data.get("Skeletons") is not None:
-            armatures = 1#len(root_data.get('Skeletons', [])) # making it '1' to work with the binary masking
+
+        output_file = r"F:\\test\\import_gr2_root_data.json"
+        with open(output_file, "w+") as f:
+            json.dump(root_data, f, indent=4)
+        #with open(log_path, "w") as f:
+            #json.dump(existing_data, f, indent=2)
+            # from dos2de (norbyte's version): context.scene.ls_properties.metadata_version = collada.ColladaMetadataLoader.LSLIB_METADATA_VERSION
+            # referring to here: C:\Users\Gabriel\AppData\Roaming\Blender Foundation\Blender\5.1\extensions\user_default\io_scene_dos2de\collada.py
+            #which has all this:
+                    #self.root = et.parse(collada_path).getroot()
+                    #self.load_root_profile(context)
+                    #anim_settings = self.find_anim_settings()
+                    #self.load_mesh_profiles()
+                    #self.load_armature_profiles()
+                    #if anim_settings is not None:
+                    #    self.load_anim_profile(context, anim_settings)
+            # Wish I'd thought to just print the entire thing before I stripped it. Would have been good to have a full reference. Oh well. Will do it later today.
+
+        if ("Skeletons" in root_data and root_data.get("Skeletons") is not None):
+            armatures = 1
         if "Meshes" in root_data and root_data.get("Meshes") is not None:
-            meshes = 1#len(root_data.get('Meshes', []))
+            meshes = 1
         if "Animations" in root_data and root_data.get("Animations") is not None:
-            animations = 1#len(root_data.get('Animations', []))
-        if "GR2Tag" in root_data and root_data.get("GR2Tag") is not None: # Does this always exist? I think so. # not sure if this is useful at all  yet. Maybe for partial automation of getting matching skeletons. Need to find out how to match to LSF data.
-            extra_data.update({"GR2Tag": root_data.get('GR2Tag')})
-        extra_data.update({"Internal_Filename": root_data['FromFileName']}) # I think  this is guaranteed to always exist? Would error in the rootreader if not.
+            animations = 1
+        if "GR2Tag" in root_data and root_data.get("GR2Tag") is not None:
+            extra_data.update({"GR2Tag": root_data['GR2Tag']})
+        if "FromFileName" in root_data and root_data.get("FromFileName") is not None:
+            extra_data.update({"Internal_Filename": root_data['FromFileName']})
+
+        print(f"In get_metadata after root_data: Filename: meshes: {meshes}, armatures: {armatures}, animations: {animations}, extra_data: {extra_data}")
         return meshes, armatures, animations, extra_data
     
     else:
         print_me(printme, f"Failed to get metadata: {result.stderr}")
         return None, None, None, None
-
-def metadata_func(input_file, armaturepath=None, printme=True):
-
-    filename = get_filename_ext(input_file)[1]
-    extra_data = {}
-
-    import_exists = check_file_exists(f"metadata: {filename}", input_file)
-    if not import_exists:
-        print_me(printme, f"Import file `{filename}` does not exist. Aborting metadata check.")
-        return 00, None, filename
     
-    if input_file == armaturepath:
-        print_me(printme, "Testing armature file for metadata:")
-        skel_meshes, skel_armatures, skel_animations, extra_data = get_metadata(armaturepath, printme)
-        if (skel_meshes, skel_armatures, skel_animations) == (0, 0, 1): # this part feels repetitive with 'get_status' now it's updated. Need to fix.
-            return 4, extra_data, filename
-    
-    print_me(printme, f"Checking metadata for `{filename}`:")
-    if armaturepath is not None and not check_file_exists("metadata: armature path", armaturepath): ## Not sure if this works. 'If it says there's a path but the file doesn't exist' is the intent.
-        armaturepath = None
-        print_me(printme, f"Armature path `{filename}` is not a valid file. Ignoring.")
+def get_status(input_file, printme): # Now includes file check internally
 
-    if ".dae" in str(input_file).lower(): # check these first, don't bother running them for metadata
-        return 7, extra_data, filename
-    elif ".glb" in str(input_file).lower() or ".gltf" in str(input_file).lower():
-        return 8, extra_data, filename
+    if check_file_exists(f"metadata: {input_file}", input_file): # returns bool
 
-    def get_status(input_file, armaturepath):
-        ARM, MSH, ANIM = 1, 2, 4
+        if ".dae" in str(input_file).lower():
+            return 7, extra_data
+        if ".glb" in str(input_file).lower() or ".gltf" in str(input_file).lower():
+            return 8, extra_data
+
+        ARM, MSH, ANIM = 1, 2, 4 #### WHEN DID THIS BECOME 43??
+
         try:
+            print(f"Input file going into get metadata: {input_file}")
             meshes, armatures, animations, extra_data = get_metadata(input_file, printme)
             mesh = meshes * MSH
             armature = armatures * ARM
             anim = animations * ANIM
             content = mesh + armature + anim
 
-            if content == 4:
-                if check_file_exists("metadata: armature path", armaturepath):
-                    skel_meshes, skel_armatures, skel_animations, _ = get_metadata(armaturepath, printme)
-                    if not skel_armatures:
-                        return 42, extra_data
-                    if skel_armatures and not (skel_meshes or skel_animations):
-                        return 41, extra_data
-                    print_me(printme, "Armature file contains more than just armature. Tentatively returning; may fail.")
-                    return 43, extra_data
-                return 42, extra_data
-
+            print(meshes, armatures, animations, extra_data)
             return content, extra_data
 
         except Exception as e:
             print_me(printme, f"FAILED TO GET METADATA: {e}")
-            return 0, extra_data
+            return 0, None
+    else:
+        return 00, None
 
-    status, extra_data = get_status(input_file, armaturepath)
-    print_me(printme, f"[{filename}] STATUS {status}: {status_definitions.get(status)} \n \n")
-    return status, extra_data, filename
+def contrast_files(f1_status, f2_status, printme):
+    if f1_status == 4: ## THIS PART has to happen after both have gone through the status check. 
+        if f2_status == 1:
+            return 41
+        if f2_status in [3,5,7]:
+            print_me(printme, "Armature file contains more than just armature. Tentatively returning; may fail.")
+            return 43
+        print(f"No valid armature file or valid armature data: {armaturepath}")
+        return 42
+    else:
+        return 00
+
+def check_if_anim_arma_pair(input_file, armaturepath=None, printme=True): 
+    # need to redo this. 
+    #need to get basic status of both if >1, /then/ decide armature role etc. not at the first found test subject.
+    contrast_1 = contrast_2 = "no contrast; not enough viable files."
+    f2_extra_data = f1_extra_data = {}
+    print("Top of `check if anim arma pair`")
+    print_me(printme, f"Checking metadata for `{input_file}` and `{armaturepath}`:")
+    f1_status, f1_extra_data = get_status(input_file, True)
+    print(f"Still here 1; {f1_status}")
+    f2_status, f2_extra_data = get_status(armaturepath, True)
+    print(f"Still here 2: {f2_status}")
+    if f1_status not in null_status and f2_status not in null_status:
+        contrast_1 = contrast_files(f1_status, f2_status, True)
+        print(f"Testing File 1 with F2 as armature: {contrast_1}")
+        contrast_2 = contrast_files(f2_status, f1_status, True)
+        print(f"Testing File 2 with F1 as armature: {contrast_2}")
+
+    filename = get_filename_ext(input_file)[1]
+    
+    arma_filename = get_filename_ext(armaturepath)[1]
+
+    print_me(printme, f"[{filename}] STATUS {f1_status}: {status_definitions.get(f1_status)}, Contrast1: {contrast_1} \n \n")
+    print_me(printme, f"[{arma_filename}] STATUS {f2_status}: {status_definitions.get(f2_status)}, Contrast2: {contrast_2} \n \n")
+    return f1_status, f1_extra_data, filename
 
 def mass_metadata(input_file_list):
 
     metadata_dict = {}
-
     for file in input_file_list:
-        # maybe check filetype first and assign status 6 immediately. Should do that.
         if "gr2" not in file.lower():
             _, filename, _ = get_filename_ext(file)
-            metadata_dict[filename] = {"local_file": file, "status": f"{6}: {status_definitions.get(6)}"}
+            metadata_dict[filename] = {"local_file": file, "status": f"6: {status_definitions.get(6)}"}
         else:
-            status, extra_data, filename = metadata_func(file, armaturepath, printme = False) # now prints internally, don't need to print the prologue.
+            status, extra_data, filename = check_if_anim_arma_pair(file, armaturepath, printme = False)
             metadata_dict[filename] = {"local_file": file, "status": f"{status}: {status_definitions.get(status)}", "GR2Tag": extra_data.get("GR2Tag"), "Internal_Filename": extra_data.get("Internal_Filename")} #moved these here, move back to separate 'if extra_data' if it causes errors.
-        #if extra_data:
-        #    metadata_dict[filename].update({
-         #   "GR2Tag": extra_data.get("GR2Tag"),
-        #    "Internal_Filename": extra_data.get("Internal_Filename")
-        #})
 
     if json_output:
         import json
         with open(jsonout_file, "w+") as f:
             json.dump(metadata_dict, f, indent=2)
         
-
 #### IMPORT HELPERS ####
-
 def conform_to_armature(filepath, armaturepath):
 
-    ## Changed this to DAE on a whim. Not sure if it's necessary or not, will test more tomorrow.
-    # Need to try making it GLB immediately, see if i can skip the middle steps.
     has_armature = [1, 3, 5, 7]
-    newfile_ext = "gr2"#changed back to GR2 so I can check the metadata. The animations to do tend to go directly from GR2 to GLTF once converted, though.
+    newfile_ext = "gr2"
     if armaturepath != None:
         try:
-            arma_status, _, _ = metadata_func(armaturepath, armaturepath, printme=True) # Should use logging levels instead of printme but it'll do for now.
+            arma_status, _, _ = check_if_anim_arma_pair(armaturepath, armaturepath, printme=True) # Should use logging levels instead of printme but it'll do for now.
             if arma_status not in has_armature:#[1, 4, 5, 6]:
                 print("Provided armature path does not contain a skeleton. Aborting conform process.")
                 return None
@@ -264,7 +271,6 @@ def conform_to_armature(filepath, armaturepath):
         except Exception as e:
             print(f"Failed to generate GR2 with new skeleton. Returning early. Reason: {e}")
             return None
-
     return f"{temppath}.{newfile_ext}"
 
 def convert_to_DAE(filepath, temppath):
@@ -285,9 +291,7 @@ def convert_to_GLB(temppath, fromtype):
     else:
         temppath = str(temppath)
     temppath2 = str(get_temppath())
-
     fromtype = fromtype.replace(".", "")
-    
     #print(f"Divine CLI command for {fromtype.upper()} to GLB conversion:")
     #print(f'"{divineexe}" --loglevel all -g bg3 -s "{temppath}" -d "{temppath2}" -i {fromtype} -o glb -a convert-model -e flip-uvs')
     #print()
@@ -323,24 +327,35 @@ def import_glb(filename, directory, existing_objects):
         return None
     return new_objects
 
-def setup_for_import(filepath):
+def setup_for_import(filepath, settings, armaturepath):
     
     _, filename, _ = get_filename_ext(filepath)
     collection = None
 
-    if specified_collection:
+    specified_collection = settings.get("collection_name")
+    reuse_existing_collection = settings.get("reuse_existing_collection")
+    
+    if specified_collection and specified_collection != "":
         trimmed_name = specified_collection
     else:
-        trimmed_name = filename.split(".")[0]
+        import_type = settings.get("import_type")
+        if import_type == "bulk_anim":
+            armature_name = get_filename_ext(armaturepath)[1]
+            trimmed_name = armature_name.split(".")[0]
+            specified_collection = True # keeps them all in one collection
+        else:
+            trimmed_name = filename.split(".")[0]
 
-    if use_existing_collection: ## Add the option to import to selected/active collection, and/or named collection in the UI once it exists.
+    if specified_collection or reuse_existing_collection: ## Add the option to import to selected/active collection, and/or named collection in the UI once it exists.
         test = bpy.data.collections.get(trimmed_name)
         if test:
-            #print(f"There is already a collection with this name: {trimmed_name}.")
+            print(f"Existing collection found with this name: {trimmed_name}.")
             collection = test
 
-    if not use_existing_collection or not collection:
+    #print(f"Collection: {collection}")
+    if not collection or (collection and not reuse_existing_collection):
         collection = bpy.data.collections.new(trimmed_name)
+        #print(f"Collection: {collection}")
 
     try:
         bpy.context.scene.collection.children.link(collection)  ## NOTE: Will fail if the collection is excluded from the view layer. Even if it passed test.
@@ -357,30 +372,25 @@ def setup_for_import(filepath):
 
     layer_collection = bpy.context.view_layer.layer_collection.children[collection.name] ## this will fail if the collection isn't linked to the view layer.
     bpy.context.view_layer.active_layer_collection = layer_collection
-
     existing_objects = set(bpy.context.scene.objects)
 
     return existing_objects
 
 def attempt_conversion(filepath, armaturepath):
-
 # I want to add a json output for conversion. The start files/types, successes and failures. Should make it far clearer to find the patterns of what succeeds/fails if I don't have to remember across runs.
     anim_or_static = "static"
-    
     has_anim = [4,41,42,43,5,7]
     
-    print("\n" *20)
+    print("\n" *6)
     print("Beginning import process...")
     print()
 
-    status, _, filename = metadata_func(filepath, armaturepath, printme = True)
-    ext = get_filename_ext(filepath)[2] # check if this works.
-    # if not, go back to // _, filename, ext = get_filename_ext(filepath)
-
+    status, _, filename = check_if_anim_arma_pair(filepath, armaturepath, printme = True)
+    ext = get_filename_ext(filepath)[2]
     if status in has_anim:
         anim_or_static = "animation"
     
-    if status in [0, 00, 42] or status is None:
+    if status in [0, 00, 42, None]:
         print(f"[{status} : {status_definitions.get(status)}]. Cannot proceed with import.")
         return None, None
 
@@ -411,7 +421,7 @@ def attempt_conversion(filepath, armaturepath):
         combined_path = conform_to_armature(filepath, armaturepath)
         if check_file_exists("attemptimport: combined path", combined_path):
             print("Combined GR2 file created successfully. Updated metadata check:")
-            new_status, _, _ = metadata_func(combined_path, armaturepath, printme = True)
+            new_status, _, _ = check_if_anim_arma_pair(combined_path, armaturepath, printme = True)
             if new_status in [5, 7]:
                 print("Attempt direct conversion of armature + anim to glb.")
                 glb_path = convert_to_GLB(combined_path, get_filename_ext(combined_path)[2])
@@ -419,6 +429,11 @@ def attempt_conversion(filepath, armaturepath):
             print("Merging of animation + armature did not result in a file with both animation and armature. Will fail.")
             return glb_path, anim_or_static
         return None, anim_or_static
+
+    if status == 5:
+        print("GR2 armature with animation")
+        glb_path = convert_to_GLB(filepath, get_filename_ext(filepath)[2])
+        return glb_path, anim_or_static
 
     if status == 6:
         if "dae" in ext:
@@ -441,7 +456,7 @@ def attempt_conversion(filepath, armaturepath):
     else:
         print(f"Anything left at the end of this function: {status}, {filename}")
         
-def cleanup(new_objects, status):
+def cleanup(new_objects, status, anim_filename, settings):
     
     # Delete LOD objects ending with _LOD\d+
     lod_pattern = re.compile(r'.*_LOD\d+')
@@ -474,17 +489,41 @@ def cleanup(new_objects, status):
     for obj in non_lod:
         if obj.type == "ARMATURE":
             armature_list.append(obj) # changed back # changed to obj.name from obj.
-            for bone in obj.pose.bones:
-                if bone.custom_shape or custom_bones_on:
+            if settings.get("custom_bones"):
+                new_custom_name = settings.get("custom_bone_obj")
+                all_obj = set(bpy.context.scene.objects)
+                for test in all_obj: ### this is a terrible way to do this.
+                    if test.name == new_custom_name:
+                        primary = test
+                    #this is going to be a next iter, isn't it. It always is.
+                for bone in obj.pose.bones:
                     if bone.custom_shape:
+                        scale_custom_bone = settings.get("scale_custom_bone")
+                        if scale_custom_bone:
+                            bone.use_custom_shape_bone_size = True
+
                         old_ico = bone.custom_shape
                         bone.custom_shape = primary
                         if old_ico != primary:
                             oldicos.add(old_ico)
                     else:
                         bone.custom_shape = primary
- 
+            else:
+                for bone in obj.pose.bones:
+                    if bone.custom_shape:
+                        bone.custom_shape = None # likely won't work. Oh, it does.
+                ## NOTE: This does not remove the original created ico or its collection. Those area remnant of the native importer.
+                # might delete them later, if they're not used by anything after this point?
+                # will look into it. See if they're data linked to anything, if not maybe delete, and then if collection empty, remove it too.
             def fix_bone_orientation():
+
+                ## NOTE: There are certain dummy bones that don't flip orientation (eg Dummy_R_Foot_01/L, Dummy_R_KneeFX_02/L, Dummy_L_Foot_02/R.) Might be able to use that as a stable reference if it holds across armatures.
+                # Also: DummyLFoot02 and DummyLFoot01 both start from the same relative position, but DummyLFoot02 is twice as long, with its tail far above  the joint. DummyLFoot01 meets the joint neatly.
+                # Not sure why and don't know what it means.
+
+                donotmove_bones = ["Root_M", ] # just a hardcoded list of bones that should not reorient.
+
+                # will save an armature from the 4.3.2 native importer and compare like bones. See where it's still not quite right.
 
                 context = bpy.context.object.data.edit_bones
                 bpy.ops.object.mode_set(mode='EDIT') # just for set to edit mode, whether it already was or not doesn't seem to error it.
@@ -515,7 +554,18 @@ def cleanup(new_objects, status):
                     if not parent:
                         #print(f"No parent for {bone.name}.")
                         pass
+                    elif parent in donotmove_bones: # think this should work?
+                        continue
                     else:
+                        siblings = bone_dict.get(parent).get("children")
+                        print(f"Number of siblings: {len(siblings)}")
+
+                        if len(siblings) > 1:
+                            print("Too many kids.")
+                            for sib in siblings:
+                                print(sib)
+                            continue
+
                         child = bone.name # could skip this and just keep typing bone.name but I need the clarity for now.
                         #print(f"parent: {parent}, child: {child}")
                         children.add(child)
@@ -533,9 +583,11 @@ def cleanup(new_objects, status):
                     if entry.get("new_tail_pos") == entry.get("tail_pos"):
                         no_movement.append(name)
                 
-                if counter != len(no_movement):
+                if counter != len(no_movement): ## this whole part is silly. 
                     print(f"Total of {counter} bones, {len(no_movement)} did not move. `({no_movement})`")   
                     for bone in no_movement:
+                        if bone in donotmove_bones:
+                            continue
                         parent = bone_dict[bone].get("parent")
                         if not parent:
                             continue
@@ -585,12 +637,19 @@ def cleanup(new_objects, status):
                         if bone:
                             bone.roll = roll
 
-            if status != "animation": # for now, just don't apply the fix to animation. I did used to ahve to retarget because I couldn't fix it automatically. Maybe still true.
-                fix_bone_orientation()
+            #print(f"Object name: {obj.name}")
+            obj.name = anim_filename[1].split(".")[0]
+            #print(f"Anim filename: {anim_filename}")
+            #print(f"New object name: {obj.name}")
 
-    for ico in oldicos: # delete here to stop structRNA errors when they're deleted before all objs have been observes.
-        bpy.data.objects.remove(ico,do_unlink=True)  ### Only needed if there's more that one option. Otherwise it deletes the one it just created and errors.
-      
+            if status != "animation": # for now, just don't apply the fix to animation. I did used to have to retarget because I couldn't fix it automatically. Maybe still true.
+                if settings.get("fix_bones"):
+                    fix_bone_orientation()
+            #fix_bone_orientation()
+
+    for ico in oldicos:
+        bpy.data.objects.remove(ico,do_unlink=True)
+
     bpy.ops.object.mode_set(mode='OBJECT')
 
     for item in armature_list:
@@ -598,13 +657,14 @@ def cleanup(new_objects, status):
 
     return armature_list
 
-def set_up_bulk_convert(armaturepath, anim_dir):
+def set_up_bulk_convert(armaturepath, anim_dir, settings):
 
     # testing armature: F:\BG3 Extract PAKs\PAKs\Models\Public\Shared\Assets\Characters\_Anims\_Creatures\Intellect_Devourer\INTDEV_Base.GR2
     # testing dir: #  F:\test\gltf_tests\raw_intdev_anims
     print("Warning: May take a while if there are many files to convert.")
-    imported_anims = [f"Armature used: {get_filename_ext(armaturepath)[1]}"]
-    #get filelist from dir
+    # HERE check the animation file is real first
+    if check_file_exists("testing animation file before bulk conversion", armaturepath):
+        imported_anims = [f"Armature used: {get_filename_ext(armaturepath)[1]}"]
     import os
     test_list = os.listdir(anim_dir)
     print(f"Test list: {test_list}")
@@ -612,13 +672,13 @@ def set_up_bulk_convert(armaturepath, anim_dir):
         print(f"Anim in test list: {anim}")
         filepath = anim_dir + "\\" + anim
         if check_file_exists("testing animation file before bulk conversion", filepath):
-            anim_file = import_process(filepath, armaturepath)
+            anim_file = import_process(filepath, armaturepath, settings)
             imported_anims.append(anim_file)
         else:
             print("No animation file(s) found.")
     return imported_anims
 
-def import_process(file_to_import, armaturepath):
+def import_process(file_to_import, armaturepath, settings): # going to move the 
 
     print(f"About to start conversion: file_to_import = {file_to_import}, armaturepath = {armaturepath}")
     if file_to_import is not None:
@@ -626,11 +686,16 @@ def import_process(file_to_import, armaturepath):
         converted, status = attempt_conversion(file_to_import, armaturepath)
         if converted:
             directory, filename, _ = get_filename_ext(converted)
-            existing_objects = setup_for_import(file_to_import)
+            existing_objects = setup_for_import(file_to_import, settings, armaturepath)
             imported = import_glb(filename, directory, existing_objects)
 
             if imported:
-                imported_files = cleanup(imported, status)
+                anim_filename =  get_filename_ext(file_to_import) # could do this inline but useful to see clearly that it's the anim skel.
+                imported_files = cleanup(imported, status, anim_filename, settings)
+
+                ### Add custom properties with import settings (y-up, flip mesh, etc etc) to armature/mesh
+                # see 'paper trailing' here: https://github.com/SWTOR-Slicers/Granny2-Plug-In-Blender-2.8x
+
                 print("Import successful.")
                 return imported_files
             else:
@@ -643,77 +708,121 @@ def import_process(file_to_import, armaturepath):
         print("File to import is None. Exiting.")
         return "No file to import."
 
-def assign_files(file):
-    print(f"File to import: {file}")
-    if metadata and ".gr2" in str(file).lower():
-        print("Checking metadata first.")
-        status, _, _ = metadata_func(file)
-        return status
-    return 6
+## I don't need this here but the syntax is useful:
+# # metadata == True if import_type == "metadata"
+# # ==     metadata = (import_type == "metadata")
+# extendable: 
+#       metadata = import_type in ("metadata", "meta", "md")
 
 def main(file_1, file_2, settings):
 
     f1_status = f2_status = file_to_import = armaturepath = None
-
+    print(f"Start of main: {file_1}, {file_2}")
     import_type = settings.get("import_type")
 
     if import_type == "bulk_anim":
         print("Bulk Anim Mode.")
-        if assign_files(file_1) == 1:
+        f1_status, _ = get_status(file_1, True)
+        if f1_status == 1:
             armaturepath = file_1
-        elif assign_files(file_1) in has_armature:
+        elif f1_status in has_armature:
             print("Armature file is not just armature: may fail.")
             armaturepath = file_1
         else:
             print("Armature file does not contain a viable armature. Returning.")
             return f"No viable armature in {file_1}"
+        
         anim_dir = file_2
     #    settings = {"collection_name":props.collection_name_override, "import_type":props.import_type, 
     #                "custom_bones":props.use_custom_bone_obj, "custom_bone_obj":props.custom_bone_obj, 
     #                "remove_all_temp":props.remove_temp, "keep_final":props.keep_final}
-        file_list = set_up_bulk_convert(armaturepath, anim_dir)
+        file_list = set_up_bulk_convert(armaturepath, anim_dir, settings)
         return file_list
 
     else:
-        f1_status = assign_files(file_1)
+        #def get_viable(file_1, file_2):
+        f1_status, _ = get_status(file_1, True) ### this is so dumb. I need to step away for a bit, I've got like 4 functions now for 'is this a skeleton' when it was already working. Need to step away, maybe try to sleep at some point, and come back.
         print(f"f1 status: {f1_status}")
-        f2_status = assign_files(file_2)
+        f2_status, _ = get_status(file_2, True)
         print(f"f2 status: {f2_status}")
-    #allows for more varied armature files, not sure whether it's good or not.
-        if file_1 in has_armature and file_2 not in has_armature and f2_status not in [0, 00, None]: # has armature
+        if f1_status in null_status and f2_status in null_status:
+            return "No viable files to import. Sorry."
+        #    if f1_status in null_status:
+        #        if f2_status in null_status:
+        #            return False, False
+        #        return False, True
+        #    if f2_status in null_status:
+        #        return True, False
+        #    return True, True ## I think this is right.##
+
+        #viable_1, viable_2 = get_viable(file_1, file_2)
+
+        #if viable_1 and not viable_2:
+        #    viable_1 = file_to_import,
+        #    armaturepath = None
+        #elif viable_2 and not viable_1:
+        #    viable_2 = file_to_import,
+        #    armaturepath = None
+#ARM, MSH, ANIM = 1, 2, 4
+
+### WHAT TO DO WITH 'ANIM+ARMA', // 'ARMA'? Do I overwrite the arma with the new one? Maybe that's an option. 'If separate armature is supplied, will always try to add it.
+#Really thuogh I should just go back to 'supplementary armature in the second slot. God this is not worth it.
+        if f1_status in has_armature and f2_status not in has_armature and f2_status not in null_status: # has armature (Should this be checking that neither f1 or f2 are nulled?)
             file_to_import = file_2
             armaturepath = file_1
-        elif file_2 in has_armature and file_1 not in has_armature and f1_status not in [0, 00, None]: # has armature
+            print("File 1 has armature, file 2 doesn't.")
+        elif f2_status in has_armature and f1_status not in has_armature and f1_status not in null_status: # has armature
             file_to_import = file_1
             armaturepath = file_2
-        if file_to_import is None and armaturepath != None:
-            file_to_import = armaturepath
-        else:
+            print("File 2 has armature, file 1 doesn't.")
+        if f1_status in null_status:
+            print(f"File one doesn't exist: {file_1}")
+            if f2_status not in null_status:
+                print(f"But file 2 does: {file_2}")
+                file_to_import = file_2
+                #print("File 1 doesn't exist, armature does.")
+        
+        elif f1_status not in null_status:
             file_to_import = file_1
+            armaturepath = file_2 # just the default of file1,
+            print("Nothing else applies, assigning file 1 as file to import.")
 
-        imported_files = import_process(file_to_import, armaturepath)
+        print(f"About to enter import process. File to import: {file_to_import}. Armaturepath: {armaturepath}")
+        imported_files = import_process(file_to_import, armaturepath, settings)
         file_list = [get_filename_ext(file_to_import)[1], imported_files]
         return file_list
 
 def run(mode, inputs, settings=None):
     metadata_collection = []
-    idx = 1
+
+    initial_setup(mode, settings)
+    armaturepath = None
     if mode == "metadata":
+        idx = 1
         for file in inputs:
+            print(f"File in inputs: {file}")
+            status, _ = get_status(file, True)
+            if status == 1:
+                armaturepath = file
+                print(f"[run, after initial setup] Armature in {file}")
             if "Secondary file (if needed)" in file:
                 metadata_collection.append("[File 2: -- No secondary file --]")
+                print("[after initial setup] No secondary file")
                 continue
-            if ".gr2" not in str(file).lower() and file.strip() != "":
+            if ".gr2" not in str(file).lower() and file.strip() != "": # change this to check if it's a file at all maybe. simple regex? Maybe not worth it.
                 print(f"[WARN] Cannot get metadata for non-GR2 files ({str(file)}).")
                 metadata_collection.append(f"[File {idx}: [{get_filename_ext(file)[1]}] STATUS 6: {status_definitions.get(6)}] \n \n")
-                idx += 1
+                idx += 1 
                 continue
             if file.strip() == "":
                 print(f"No file in fileline {idx}")
                 metadata_collection.append(f"[File {idx}: STATUS 0: No Filepath] \n \n")
                 idx += 1
                 continue
-            status, _, filename = metadata_func(file_to_import, None, False)
+            if status in null_status:
+                print(f"File is null status: {file}, {status}")
+            print(f"About to go to metadata func from run for {file}")
+            status, _, filename = check_if_anim_arma_pair(file, armaturepath, True) #temporarily true for troubleshooting.
             metadata_collection.append(f"[File {idx}: [{filename}] STATUS {status}: {status_definitions.get(status)}] \n \n")
             idx += 1
 
@@ -722,7 +831,8 @@ def run(mode, inputs, settings=None):
         return metadata_collection
 
     if mode == "import":
-        [file_1, file_2] = inputs
+        print(f"Mode import. Inputs: {inputs}, type: {type(inputs)}")
+        file_1, file_2 = inputs # which one?
+        #[file_1, file_2] = inputs
         file_list = main(file_1, file_2, settings)
         return file_list
-